Move checksum functions to car modules (#2541)
* Move checksum helpers into can modules

* cleanup

* one more
