+++ opendbc/car/chrysler/chryslercan.py
@@ -1,4 +1,5 @@
 from opendbc.car import structs
+from opendbc.car.crc import CRC8J1850
 from opendbc.car.chrysler.values import RAM_CARS
 
 GearShifter = structs.CarState.GearShifter
@@ -69,3 +70,43 @@ def create_cruise_buttons(packer, frame, bus, cancel=False, resume=False):
     "COUNTER": frame % 0x10,
   }
   return packer.make_can_msg("CRUISE_BUTTONS", bus, values)
+
+
+def chrysler_checksum(address: int, sig, d: bytearray) -> int:
+  checksum = 0xFF
+  for j in range(len(d) - 1):
+    curr = d[j]
+    shift = 0x80
+    for _ in range(8):
+      bit_sum = curr & shift
+      temp_chk = checksum & 0x80
+      if bit_sum:
+        bit_sum = 0x1C
+        if temp_chk:
+          bit_sum = 1
+        checksum = (checksum << 1) & 0xFF
+        temp_chk = checksum | 1
+        bit_sum ^= temp_chk
+      else:
+        if temp_chk:
+          bit_sum = 0x1D
+        checksum = (checksum << 1) & 0xFF
+        bit_sum ^= checksum
+      checksum = bit_sum & 0xFF
+      shift >>= 1
+  return (~checksum) & 0xFF
+
+
+def fca_giorgio_checksum(address: int, sig, d: bytearray) -> int:
+  crc = 0
+  for i in range(len(d) - 1):
+    crc ^= d[i]
+    crc = CRC8J1850[crc]
+  if address == 0xDE:
+    return crc ^ 0x10
+  elif address == 0x106:
+    return crc ^ 0xF6
+  elif address == 0x122:
+    return crc ^ 0xF1
+  else:
+    return crc ^ 0x0A
