+++ opendbc/car/honda/hondacan.py
@@ -213,3 +213,21 @@ def spam_buttons_command(packer, CAN, button_val, car_fingerprint):
   # send buttons to camera on radarless (camera does ACC) cars
   bus = CAN.camera if car_fingerprint in HONDA_BOSCH_RADARLESS else CAN.pt
   return packer.make_can_msg("SCM_BUTTONS", bus, values)
+
+
+def honda_checksum(address: int, sig, d: bytearray) -> int:
+  s = 0
+  extended = address > 0x7FF
+  addr = address
+  while addr:
+    s += addr & 0xF
+    addr >>= 4
+  for i in range(len(d)):
+    x = d[i]
+    if i == len(d) - 1:
+      x >>= 4
+    s += (x & 0xF) + (x >> 4)
+  s = 8 - s
+  if extended:
+    s += 3
+  return s & 0xF
