+++ opendbc/car/tesla/teslacan.py
@@ -1,5 +1,4 @@
 from opendbc.car.common.conversions import Conversions as CV
-from opendbc.car.interfaces import V_CRUISE_MAX
 from opendbc.car.tesla.values import CANBUS, CarControllerParams
 
 
@@ -17,6 +16,8 @@ class TeslaCAN:
     return self.packer.make_can_msg("DAS_steeringControl", CANBUS.party, values)
 
   def create_longitudinal_command(self, acc_state, accel, counter, v_ego, active):
+    from opendbc.car.interfaces import V_CRUISE_MAX
+
     set_speed = max(v_ego * CV.MS_TO_KPH, 0)
     if active:
       # TODO: this causes jerking after gas override when above set speed
@@ -40,3 +41,12 @@ class TeslaCAN:
     }
 
     return self.packer.make_can_msg("APS_eacMonitor", CANBUS.party, values)
+
+
+def tesla_checksum(address: int, sig, d: bytearray) -> int:
+  checksum = (address & 0xFF) + ((address >> 8) & 0xFF)
+  checksum_byte = sig.start_bit // 8
+  for i in range(len(d)):
+    if i != checksum_byte:
+      checksum += d[i]
+  return checksum & 0xFF
